ListDictionary moved from the Catel.Cache to the Catel.Collections. Improved xml documrntation
