+++ src/Catel.Core/Catel.Core.Shared/Data/CatelTypeInfo.cs
@@ -14,6 +14,7 @@ namespace Catel.Data
     using Caching;
     using Catel.Logging;
     using Catel.Reflection;
+    using Collections;
 
     /// <summary>
     /// Class containing all information about a Catel type (such as properties).
